Updated Footer - the footer is now reactive to the current path, so it only shows options besides the current site. Updated header - the header has now instead of tool two new titles. The margin has been adapted accordingly. The plot and download feature has been split apart. Those are now separate pages. The new home page is now a landing page, highlighting those two pages. The images there still need to be changed, a news box needs to be added. The download and plot page are not finished yet either but work in progress.
